refactor: Refactor AppBar component and extract Logo component
